rename user service to profile service
